backend/requirements.txt
FastAPI
uvicorn
sqlalchemy
mysqlclient
pydantic
python-dotenv
httpx
